// database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      users: {
        Row: {
          clerk_id: string
          email: string
          level: number
          name: string
          onboarded_at: string
        }
        Insert: {
          clerk_id: string
          email: string
          level: number
          name: string
          onboarded_at?: string
        }
        Update: {
          clerk_id?: string
          email?: string
          level?: number
          name?: string
          onboarded_at?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}